% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_utils.R
\name{filter_items_by_property}
\alias{filter_items_by_property}
\title{Filter STAC items by property}
\usage{
filter_items_by_property(items, property, value)
}
\arguments{
\item{items}{List of STAC items}

\item{property}{Property name}

\item{value}{Value to match}
}
\value{
Filtered list of items
}
\description{
Filter STAC items by property
}
\examples{
items <- load_stac_items()
filter_items_by_property(items, "wri_domain", "livelihoods")
}
