% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_utils.R
\name{get_asset_urls}
\alias{get_asset_urls}
\title{Get asset URLs from a STAC item}
\usage{
get_asset_urls(item, asset_type = NULL)
}
\arguments{
\item{item}{A STAC item}

\item{asset_type}{Optional asset type (e.g., "data")}
}
\value{
Character vector of asset URLs
}
\description{
Get asset URLs from a STAC item
}
\examples{
items <- load_stac_items()
get_asset_urls(items[[1]])
}
