% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_utils.R
\name{get_data_types}
\alias{get_data_types}
\title{(Deprecated) Extract unique data types from STAC items}
\usage{
get_data_types(items)
}
\arguments{
\item{items}{List of STAC items}
}
\value{
Character vector of unique data types
}
\description{
(Deprecated) Extract unique data types from STAC items
}
\examples{
items <- load_stac_items()
list_stac_properties(items)$data_type
}
