% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_utils.R
\name{get_domains}
\alias{get_domains}
\title{(Deprecated) Extract unique domains from STAC items}
\usage{
get_domains(items)
}
\arguments{
\item{items}{List of STAC items}
}
\value{
Character vector of unique domains
}
\description{
(Deprecated) Extract unique domains from STAC items
}
\examples{
items <- load_stac_items()
list_stac_properties(items)$wri_domain
}
