% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_utils.R
\name{get_layer_types}
\alias{get_layer_types}
\title{(Deprecated) Extract unique layer types from STAC items}
\usage{
get_layer_types(items)
}
\arguments{
\item{items}{List of STAC items}
}
\value{
Character vector of unique layer types
}
\description{
(Deprecated) Extract unique layer types from STAC items
}
\examples{
items <- load_stac_items()
list_stac_properties(items)$wri_layer_type
}
