% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_domains.R
\name{list_domains}
\alias{list_domains}
\title{List available WRI domains}
\usage{
list_domains()
}
\value{
A character vector of domain names
}
\description{
List available WRI domains
}
\examples{
# See all available domains
list_domains()

# Check if a domain is valid
"water" \%in\% list_domains()

# Loop through domains
for (domain in list_domains()) {
  print(domain)
}
}
