% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{list_layer_types}
\alias{list_layer_types}
\title{List available layer types}
\usage{
list_layer_types()
}
\value{
Character vector of layer type names
}
\description{
List available layer types
}
\examples{
list_layer_types()
}
