% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_utils.R
\name{list_stac_item_files}
\alias{list_stac_item_files}
\title{List all STAC item files}
\usage{
list_stac_item_files()
}
\value{
List of item file paths
}
\description{
List all STAC item files
}
\examples{
files <- list_stac_item_files()
head(files)
}
