% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_utils.R
\name{list_stac_properties}
\alias{list_stac_properties}
\title{List all unique property names and values in STAC items}
\usage{
list_stac_properties(items)
}
\arguments{
\item{items}{List of STAC items}
}
\value{
Named list of unique values for each property
}
\description{
List all unique property names and values in STAC items
}
\examples{
items <- load_stac_items()
list_stac_properties(items)
}
