% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_utils.R
\name{load_stac_items}
\alias{load_stac_items}
\title{Load all STAC items}
\usage{
load_stac_items()
}
\value{
List of STAC item objects
}
\description{
Load all STAC items
}
\examples{
items <- load_stac_items()
length(items)
}
