% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_utils.R
\name{summarize_stac_items}
\alias{summarize_stac_items}
\title{Summarize all STAC items in a data frame}
\usage{
summarize_stac_items(items, properties = NULL)
}
\arguments{
\item{items}{List of STAC items}

\item{properties}{Character vector of property names to include (default: all)}
}
\value{
Data frame summary
}
\description{
Summarize all STAC items in a data frame
}
\examples{
items <- load_stac_items()
summarize_stac_items(items, c("id", "wri_domain", "data_type"))
}
