% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_config.R
\name{get_stac_catalog_path}
\alias{get_stac_catalog_path}
\title{Get the local STAC catalog path}
\usage{
get_stac_catalog_path()
}
\value{
Character path to the STAC catalog
}
\description{
Get the local STAC catalog path
}
